+++ lib/season_statistics.rb
@@ -78,32 +78,6 @@ module SeasonStatistics
     end
   end
 
-  def season_summary(season, team_id) #tested line 287
-    season = season.to_i
-    team_id = team_id.to_i
-    summary = {:preseason => {}, :regular_season => {}}
-    pre_wins_percent = wins_percentage(season, "P")
-    reg_wins_percent = wins_percentage(season, "R")
-      if pre_wins_percent.include?(team_id)
-        summary[:preseason][:win_percentage] = pre_wins_percent.find {|team,wins| team == team_id}[1].round(2)
-      else summary[:preseason][:win_percentage] = 0.0
-      end
-      if reg_wins_percent.include?(team_id)
-        summary[:regular_season][:win_percentage] = reg_wins_percent.find {|team,wins| team == team_id}[1].round(2)
-      else summary[:preseason][:win_percentage] = 0.0
-      end
-    preseason_games = game_by_type(season, "P")
-    p_goals_scored = away_goals_scored(preseason_games, team_id) + home_goals_scored(preseason_games, team_id)
-    summary[:preseason][:goals_scored] = p_goals_scored
-    p_goals_allowed = away_goals_allowed(preseason_games, team_id) + home_goals_allowed(preseason_games, team_id)
-    summary[:preseason][:goals_against] = p_goals_allowed
-    regular_season_games = game_by_type(season, "R")
-    r_goals_scored = away_goals_scored(regular_season_games, team_id) + home_goals_scored(regular_season_games, team_id)
-    summary[:regular_season][:goals_scored] = r_goals_scored
-    r_goals_allowed = away_goals_allowed(regular_season_games, team_id) + home_goals_allowed(regular_season_games, team_id)
-    summary[:regular_season][:goals_against] = r_goals_allowed
-    return summary
-  end
