+++ lib/seasonal_summary.rb
@@ -14,6 +14,30 @@ module SeasonalSummary
     end.compact.uniq
   end
 
+  def away_goals_allowed(games, team_id) #tested line 269
+    away_goals_allowed = 0
+    @games_data.each do |game|
+      if games.include?(game.game_id)
+        if game.away_team_id == team_id
+          away_goals_allowed += game.home_goals
+        end
+      end
+    end
+    return away_goals_allowed
+  end
+
+  def home_goals_scored(games, team_id) #tested line 275
+    home_goals_scored = 0
+    @games_data.each do |game|
+      if games.include?(game.game_id)
+        if game.home_team_id == team_id
+          home_goals_scored += game.home_goals
+        end
+      end
+    end
+    return home_goals_scored
+  end
+
     def create_seasonal_summary_hash(seasons)
       seasonal_summary = Hash.new(0)
       seasons.each do |season|
@@ -76,6 +100,33 @@ module SeasonalSummary
           end
       end
       return seasonal_summary
+    end
+
+    def season_summary(season, team_id) #tested line 287
+      season = season.to_i
+      team_id = team_id.to_i
+      summary = {:preseason => {}, :regular_season => {}}
+      pre_wins_percent = wins_percentage(season, "P")
+      reg_wins_percent = wins_percentage(season, "R")
+        if pre_wins_percent.include?(team_id)
+          summary[:preseason][:win_percentage] = pre_wins_percent.find {|team,wins| team == team_id}[1].round(2)
+        else summary[:preseason][:win_percentage] = 0.0
+        end
+        if reg_wins_percent.include?(team_id)
+          summary[:regular_season][:win_percentage] = reg_wins_percent.find {|team,wins| team == team_id}[1].round(2)
+        else summary[:preseason][:win_percentage] = 0.0
+        end
+      preseason_games = game_by_type(season, "P")
+      p_goals_scored = away_goals_scored(preseason_games, team_id) + home_goals_scored(preseason_games, team_id)
+      summary[:preseason][:goals_scored] = p_goals_scored
+      p_goals_allowed = away_goals_allowed(preseason_games, team_id) + home_goals_allowed(preseason_games, team_id)
+      summary[:preseason][:goals_against] = p_goals_allowed
+      regular_season_games = game_by_type(season, "R")
+      r_goals_scored = away_goals_scored(regular_season_games, team_id) + home_goals_scored(regular_season_games, team_id)
+      summary[:regular_season][:goals_scored] = r_goals_scored
+      r_goals_allowed = away_goals_allowed(regular_season_games, team_id) + home_goals_allowed(regular_season_games, team_id)
+      summary[:regular_season][:goals_against] = r_goals_allowed
+      return summary
     end
 
   end
