+++ lib/team_statistics_hm.rb
@@ -1,5 +1,10 @@
 module TeamStatisticsHM
 
+  def team_id_name(id) #tested line 132
+    team = @teams_data.find { |team| team.team_id == id }
+    return team.team_name
+  end
+
     def games_played_by_season(team_id)
       games_by_season = @games_data.group_by { |game| game.season }
         games_by_season.each do |season,games|
@@ -11,7 +16,6 @@ module TeamStatisticsHM
       return games_by_season
     end
 
-  # Helper Method
     def games_won_by_season(team_id) #tested line 299
       team_id = team_id.to_i
       games_by_season = games_played_by_season(team_id)
@@ -32,8 +36,7 @@ module TeamStatisticsHM
     end
 
 
-    # Helper Method
-      def all_team_goals(team_id) #tested line 320
+        def all_team_goals(team_id) #tested line 320
         @games_teams_stats.map do |stat|
           if stat.team_id == team_id
             stat.goals
@@ -42,8 +45,7 @@ module TeamStatisticsHM
       end
 
 
-      # Helper Method
-        def collect_home_games_opponents(team_id, hash) #tested line 335
+            def collect_home_games_opponents(team_id, hash) #tested line 335
           @games_data.each do |game|
             if game.home_team_id == team_id
               if hash.has_key?(game.away_team_id)
@@ -56,8 +58,7 @@ module TeamStatisticsHM
         return hash
         end
 
-      # Helper Method
-        def collect_away_games_opponents(team_id, hash) #tested line 341
+            def collect_away_games_opponents(team_id, hash) #tested line 341
           @games_data.each do |game|
             if game.away_team_id == team_id
               if hash.has_key?(game.home_team_id)
@@ -70,8 +71,7 @@ module TeamStatisticsHM
           return hash
         end
 
-      # Helper Method
-        def get_opponents_results(team_id, hash) #tested line 347
+            def get_opponents_results(team_id, hash) #tested line 347
           collect_home_games_opponents(team_id, hash)
           collect_away_games_opponents(team_id, hash)
           hash.each do |k,v|
@@ -84,8 +84,7 @@ module TeamStatisticsHM
           end
         end
 
-        # Helper Method
-          def get_goals_blowout(games) #tested lines 363
+                def get_goals_blowout(games) #tested lines 363
             @games_data.map do |game|
               if games.include?(game.game_id)
                 (game.away_goals - game.home_goals).abs
@@ -115,7 +114,5 @@ module TeamStatisticsHM
                 game.away_team_id == team_id && game.home_team_id == opponent && game.outcome.include?("away") || game.home_team_id == team_id && game.away_team_id == opponent && game.outcome.include?("home")
               end
             end
-
-
 
 end
